updated tasks

+++ task_2/array.c
@@ -1,7 +1,9 @@
 #include <stdio.h>
-void array_scan(short int arr[] , int size)
+#include "STD_TYPES.h"
+
+void array_scan(u16 arr[] , u16 size)
 {
-    int counter = 0;
+    u16 counter = 0;
     for (counter = 0; counter < size; counter++)
     {
         printf("Enter number [%d]: ", counter + 1);
@@ -9,9 +11,9 @@ void array_scan(short int arr[] , int size)
     }
 }
 
-void array_display(short int arr[] , int size)
+void array_display(u16 arr[] , u16 size)
 {
-    int counter = 0;
+    u16 counter = 0;
     for (counter = 0; counter < size; counter++)
     {
         printf("number [%d]: %hd\n", counter + 1, arr[counter]);
